Real comments now

+++ scanner.h
@@ -4,9 +4,8 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-// Must define this BEFORE the gap struct and prototypes
 typedef struct {
-    int8_t id;
+    int16_t id;
     uint8_t s_angle;
     uint8_t e_angle;
 	uint8_t c_angle;
@@ -15,7 +14,7 @@ typedef struct {
 } DetectedObject;
 
 typedef struct {
-    int8_t id;
+    int16_t id;
     uint8_t s_angle;
     uint8_t e_angle;
 	uint8_t c_angle;
@@ -23,15 +22,12 @@ typedef struct {
     float l_width_cm;
 } DetectedGap;
 
-// Core Scanner Functions
 int perform_advanced_sweep(DetectedObject objects[], int max_objects);
 DetectedObject find_smallest_object(DetectedObject objects[], int num_objects);
 
-// Gap Logic
 int calculate_all_gaps(DetectedObject objects[], int num_objects, DetectedGap gaps[], int max_gaps);
 DetectedGap find_best_driveable_gap(DetectedGap gaps[], int num_gaps, int robot_width_cm);
 
-// NEW: Parking Logic
 DetectedGap check_for_parking_zone(DetectedObject objects[], int num_objects);
 
 #endif
